kkkkk

+++ C0522G1_NguyenTanHuan_Module2/src/home_work/util/ReadFileUtil.java
@@ -35,7 +35,7 @@ public class ReadFileUtil {
         String[] info;
         for(String line : strings){
             info = line.split(",");
-            students.add(new Student(Integer.parseInt(info[0]), info[1], info[2], info[3], info[4], Integer.parseInt(info[5])));
+            students.add(new Student(Integer.parseInt(info[0]), info[1], info[2], info[3], info[4], Double.parseDouble(info[5])));
         }
 
         return students;
